Tweaks: Add demand-based mod loading, logging, and fixes
Demand-based mod loading makes it so that modules are only loaded into the game once they are needed on a bomb. This means that not only is the time to start the game up reduced significantly but also the game will use less memory. To enable the feature, enable both the "Demand-Based Mod Loading" and the "Disable Demand-Based Mods" feature then follow the prompt.